feat: move command registration functionality to `kc-osm`
