+++ crates/kc-osm/Cargo.toml
@@ -10,4 +10,5 @@ description = "Helper library for writing Dark Engine object script modules (OSM
 [dependencies]
 windows = { version = "0.62.2", features = ["Win32_System_Com"] }
 windows-core = "0.62.2"
+windows-sys = { version = "0.61.2", features = ["Win32_System_LibraryLoader"] }
 kc-osm-proc-macros = { version = "0.0.1", path = "../kc-osm-proc-macros" }
