+++ examples/custom_commands/Cargo.toml
@@ -9,4 +9,3 @@ crate-type=["cdylib"]
 [dependencies]
 kc-osm = { path = "../../crates/kc-osm" }
 windows-core = "0.62.2"
-windows-sys = { version = "0.61.2", features = ["Win32_System_LibraryLoader"] }
